feat: revert support requesting delegations with ICRC-29 (#50)
# Motivation

As discussed internally, we would like to revert PR #39, as it currently
ends up being unused and will likely be rewritten in the future if and
when reintroduced. Furthermore, we do not want to require consumers of
this library to install and bundle additional dependencies that are also
unused.

# Notes

Reverting this feature is technically a breaking change. However, since
the `@icp-sdk/auth` library and the removed function have not been
publicly advertised or tested yet and are therefore unlikely to be used
by anyone, we allow ourselves to remove it in a minor version. This also
avoids the need to move to `v5` and ensures alignment with
`@icp-sdk/core`, which has already been promoted.

# Changes

- Remove code and dependencies introduces in PR #39